Update index.md
deleted
This is the workshop template. Delete these lines and use it to customize your own website. If you are running a self-organized workshop or have not put in a workshop request yet, please also fill in this workshop request form to let us know about your workshop and our administrator may contact you if we need any extra information.